test(helper): factor out domain
[#180438173](https://www.pivotaltracker.com/story/show/180438173)
